Add composed cache to set token storage limit.
This was previously disabled (even if it was documented as working in
the README, as composition with `ttlcache` didn't seem to work. It
turns out `ttlcache` uses some internal data structures incompatible
with the rest of clojure.core.cache implementations.

As a result, `ttlcache` can only be composed in a certain way (it being the
first cache in the composition scheme).

Also `ttlcache` has a performance optimization that plays bad with
cache composition (namely it doesn't dissoc keys from the expiry heap
when evicting a key). This means sometimes keys appear to be in the
cache (when using cache/has?), but when we try to retrieve them, we
get `nil` (because they were not in the cache actually). It's a
trivial fix in `ttlcache` but it must be applied upstream (which is
not at the time of writing this).