Merge validate-single-key into validate-token*
Associated with that merge, changed the way the pubkeys from JWKS are
stored in the cache. It avoids looping over pubkeys that, while valid,
will not unsign the token (because it has been signed by another valid
key from the same key set). And logging their corresponding
failures (that are just noise, not real, hard failures).