Add JVM options to let AWS SDK call JAVA 9+ internal methods
The AWS SDK library makes some illegal reflections code
access. According to the devs it won't be fixed because they are
working in the v2 of the SDK [1]. JDK versions from 9 to 16 just
generated a warning, since JDK 17 the JVM blocks those illegal
accesses by default. That generates unexpected behaviour and
exceptions in the Amazonica library. By using this flag we tell the
JVM not to block them.

As far as we know the flag doesn't have any security implications. The
access is blocked by default because of maintainability reasons[2]. The
libraries shouldn't be calling to low-level Java code because is not
part of the public API and is subject to change.

We use this only in development, for the integration tests (we use AWS
Cognito in those tests).

[1] https://github.com/mcohen01/amazonica/issues/323#issuecomment-461392822
[2] http://openjdk.java.net/jeps/261