Fix `test-validate-token*` unit test
Make `test-validate-token*` unit test be strict about not accepting
symmetric encryption keys. Previously, we didn't include the symmetric
key in the set of keys to use for validation.

Hence any token signed with the symmetric key was considered
invalid (as the key was not found in the JWKS set). Now we add it to
the JWKS set and make sure a token correctly signed with that
symmetric key is still considered invalid.

Fixing the above issue also uncovered a related one. The test that
should fail because the token was signed with a key that was not in
the JWKS key set was using the symmetric key. In that case the code
fails early because we are using a symmetric key, and doesn't even try
to validate the token signature. So we get a false positive. So make
sure that we use an asymmetric key that is not in the JWKS set.