Move implementation-specific specs to its own namespace
We plan to moe all implemenation functions that are not part of the
public API to a `impl` namespace. We want to keep those functions
public to be able to test them from the tests namespace. But they are
not considered part of the public API of the library, and should not
be used by external parties (as they may change without notice, and
without maintaining backwards compatibility).