Fix spec for the 'aud' claim, to reject bad values
According to the OpenID Connect Core 1.0 standard[1], the aud claim
must be either a single case-sensitive string, or in the general case,
an array of case-sensitive strings. The current spec (and associated
pre-condition in authfn) correctly checks for the first
alternative (single string), but not the second (array of strings). We
accept any collection, with any cardinality (even an empty
collection!) and we don't check for the type of the collection
entries (they should all be strings, and only strings!)

We need to fix the spec for ::aud to correctly deal with the second
alternative.

[1] https://openid.net/specs/openid-connect-core-1_0.html#IDToken

[Re: #4]